ISSUE-311 Added required changes to TransactionEventListener to consider whether UnitOfWork should be applied for resources by default or not.   - Moved TransactionEventListener to storage-core module.
