+++ webservice/src/main/java/com/hortonworks/registries/webservice/RegistryApplication.java
@@ -17,7 +17,7 @@ package com.hortonworks.registries.webservice;
 
 import com.hortonworks.registries.common.GenericExceptionMapper;
 import com.hortonworks.registries.common.ServletFilterConfiguration;
-import com.hortonworks.registries.listeners.TransactionEventListener;
+import com.hortonworks.registries.storage.transaction.TransactionEventListener;
 import com.hortonworks.registries.storage.NOOPTransactionManager;
 import com.hortonworks.registries.storage.TransactionManager;
 import io.dropwizard.assets.AssetsBundle;
